Feat: Move ResidentCardCreation folder to the lower part of the modules folder
