+++ src/modules/ResidentCardCreation/ResidentCardSteps.tsx
@@ -3,8 +3,8 @@
 import { useCallback, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import { ResidentCardForm } from '@/components/ResidentCardCreation/Form';
-import { Steps } from '@/components/ResidentCardCreation/ResidentCardCreation.type';
+import { ResidentCardForm } from '@/modules/ResidentCardCreation/Form';
+import { Steps } from '@/modules/ResidentCardCreation/ResidentCardCreation.type';
 
 const steps: Steps[] = ['BOARDING', 'PROFILE', 'KEYWORD', 'KEYWORD_CONTENT', 'COMPLETE'];
